refactor: Centralize data handling functions
